+++ RELEASES.md
@@ -9,6 +9,7 @@
 - better decoupled `InputMap` and `ActionState`, providing an `InputMap::which_pressed` API and allowing `ActionState::update` to operate based on any `HashSet<A: Actionlike>` of pressed virtual buttons that you pass in
 - `InputMap` now uses a collected `InputStreams` struct in all of its methods, and input methods are now optional
 - `InputManagerPlugin` now works even if some input stream resources are missing
+- added the `input_pressed` method to `InputMap`, to check if a single input is pressed
 
 ### Bug fixes